backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
sqlalchemy>=2.0.23
python-docx>=1.1.0
PyPDF2>=3.0.1
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
jinja2>=3.1.2
pandas>=2.1.4
numpy>=1.26.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
